+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 2.3.0 (unreleased)
   * Add 'version' CLI command.
   * Rpush::Wpns::Notification now supports setting the 'data' attribute.
+  * ActiveRecord is now directed to the configured Rpush logger (#104).
+  * Logs are reopened when the HUP signal is received (#95).
 
 ## 2.2.0 (Oct 7, 2014)
   * Numerous command-line fixes, sorry folks!